devices_service/requirements.txt
cryptography

# Django and REST Framework
django==5.2.7
djangorestframework
djangorestframework-simplejwt

# API Documentation
drf-spectacular

# Environment Variables
python-decouple

# Database (PostgreSQL)
psycopg2-binary

# Redis / Caching / Celery
redis               
django-redis

# Code Quality / Formatting
black
flake8

# Testing
pytest
pytest-django